html/semantics/forms/the-input-element/button.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>input type button</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#button-state-(type=button)">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<input type=button id=button>

<script>
  test(function(){
    assert_false(document.getElementById('button').willValidate);
  }, "the element is barred from constraint validation");
</script>
